pyproject.toml
[project]
name = "openai-agent-sdk-mcp"
version = "0.1.0"
description = ""
authors = [
    {name = "katpyeon",email = "katpyeon@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "openai-agents (>=0.0.14,<0.0.15)",
    "gradio (>=5.29.0,<6.0.0)",
    "aiohttp (>=3.11.18,<4.0.0)",
    "fastmcp (>=2.2.6,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
]

[tool.poetry]
packages = [{include = "openai_agent_sdk_mcp", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
